+++ changelog.md
@@ -9,3 +9,4 @@ Beta
 - Made the Curse of Ymir configurable
 - Tweaks to OdmNodeEntity#getRightOrLeftOffset and the info text for blade handles
 - Compressed music disc files to reduce mod file size
+- Made swinging blade handles cause exhaustion